refactor(cli): move argo CLI into its own module
